Make the address-space query functionality optional under a `query` feature.
This enables the crate to build on platforms for which we haven't
implemented the necessary address-space query code (for example,
OpenBSD).  The crate is missing some functionality when built this way:
in particular, there is no scope-guard object that restores the old
protections for a memory region when a change goes out of scope.
However, some simple use-cases of this crate (as essentially a wrapper
for `mprotect` for a JIT engine, for example) might not need this
extended functionality, and so can build without default features (to
omit the `query` feature).